redid dismodgroups the way Alec suggested
